Calculate the diurnal variation in air temperature with Parton and Logan, 1981

Calculate the diurnal variation in air temperature. Parton WJ, Logan JA (1981) A model for diurnal variation in soil and air temperature. Agricultural Meteorology, 23, 205?216. Codes copied from APSIM Utilities.cpp

Usage

diurnalT(maxt, mint, doy, hour, latitude, A = 1.5, B = 4, C = 1)

Arguments

maxt

maximum daily temperature

mint

minimum daily temperature

doy

day of year

hour

hour from 1 to 24

latitude

latitude in radials

A

is the time lag in temperature after noon

B

is coef that controls temperature decrease at night

C

is the time lag for min temperature after sunrise

Value

A vector with diurnal air temperature

Examples

diurnalT(maxt = 20, mint = 10, doy  = 1, 
   hour = seq(from = 1, to = 23.99, by = 0.1), 
   latitude = -10, A = 1.5, B = 4, C = 1)
#>   [1] 10.60904 10.58507 10.56204 10.53991 10.51866 10.49825 10.47863 10.45979
#>   [9] 10.44169 10.42431 10.40760 10.39156 10.37615 10.36134 10.34712 10.33345
#>  [17] 10.32033 10.30772 10.29561 10.28397 10.27279 10.26205 10.25174 10.24183
#>  [25] 10.23231 10.22317 10.21438 10.20594 10.19784 10.19005 10.18257 10.17538
#>  [33] 10.16848 10.16185 10.15547 10.14935 10.14348 10.13783 10.13240 10.12719
#>  [41] 10.12218 10.11737 10.11275 10.10832 10.10405 10.09996 10.09602 10.09224
#>  [49] 10.08861 10.08512 10.03649 10.22084 10.40512 10.58926 10.77320 10.95688
#>  [57] 11.14023 11.32319 11.50571 11.68771 11.86914 12.04993 12.23003 12.40936
#>  [65] 12.58788 12.76552 12.94222 13.11792 13.29256 13.46608 13.63842 13.80953
#>  [73] 13.97934 14.14780 14.31485 14.48043 14.64449 14.80697 14.96781 15.12697
#>  [81] 15.28438 15.44000 15.59377 15.74564 15.89555 16.04346 16.18932 16.33307
#>  [89] 16.47467 16.61407 16.75122 16.88608 17.01859 17.14872 17.27642 17.40165
#>  [97] 17.52436 17.64451 17.76206 17.87698 17.98922 18.09874 18.20551 18.30949
#> [105] 18.41065 18.50894 18.60435 18.69683 18.78635 18.87289 18.95641 19.03689
#> [113] 19.11429 19.18860 19.25978 19.32782 19.39269 19.45436 19.51282 19.56804
#> [121] 19.62002 19.66872 19.71414 19.75625 19.79505 19.83052 19.86265 19.89142
#> [129] 19.91683 19.93888 19.95754 19.97282 19.98471 19.99320 19.99830 20.00000
#> [137] 19.99830 19.99320 19.98471 19.97282 19.95754 19.93888 19.91683 19.89142
#> [145] 19.86265 19.83052 19.79505 19.75625 19.71414 19.66872 19.62002 19.56804
#> [153] 19.51282 19.45436 19.39269 19.32782 19.25978 19.18860 19.11429 19.03689
#> [161] 18.95641 18.87289 18.78635 18.69683 18.60435 18.50894 18.41065 18.30949
#> [169] 18.20551 18.09874 17.98922 17.87698 17.76206 17.64451 17.52436 17.40165
#> [177] 17.27642 17.14872 17.01859 16.88608 16.75122 16.51112 16.25482 16.00861
#> [185] 15.77209 15.54488 15.32662 15.11694 14.91552 14.72203 14.53616 14.35760
#> [193] 14.18607 14.02129 13.86300 13.71094 13.56487 13.42454 13.28974 13.16024
#> [201] 13.03585 12.91635 12.80155 12.69127 12.58533 12.48357 12.38580 12.29189
#> [209] 12.20168 12.11501 12.03176 11.95178 11.87495 11.80115 11.73025 11.66214
#> [217] 11.59671 11.53386 11.47348 11.41548 11.35976 11.30624 11.25482 11.20543
#> [225] 11.15798 11.11240 11.06861 11.02654 10.98614 10.94732